% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_oscillations.R
\name{calculate_oscillations}
\alias{calculate_oscillations}
\title{Calculate Oscillations for Individual Variable Profiles}
\usage{
calculate_oscillations(x, sort = TRUE, ...)
}
\arguments{
\item{x}{an individual variable profile explainer produced with the `individual_variable_profile()` function}

\item{sort}{a logical value. If TRUE then rows are sorted along the oscillations}

\item{...}{other arguments}
}
\description{
Calculate Oscillations for Individual Variable Profiles
}
\examples{
library("DALEX")
 \dontrun{
library("randomForest")
set.seed(59)

apartments_rf_model <- randomForest(m2.price ~ construction.year + surface + floor +
      no.rooms + district, data = apartments)

explainer_rf <- explain(apartments_rf_model,
      data = apartmentsTest, y = apartmentsTest$m2.price)

apartment <- apartmentsTest[1,]

cp_rf <- individual_variable_profile(explainer_rf, apartment)
calculate_oscillations(cp_rf)
}
}
