% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ceteris_paribus.R
\name{print.individual_variable_profile_explainer}
\alias{print.individual_variable_profile_explainer}
\title{Prints Individual Variable Explainer Summary}
\usage{
\method{print}{individual_variable_profile_explainer}(x, ...)
}
\arguments{
\item{x}{an individual variable profile explainer produced with the `individual_variable_profile()` function}

\item{...}{other arguments that will be passed to `head()`}
}
\description{
Prints Individual Variable Explainer Summary
}
\examples{
library("DALEX")
 \dontrun{
library("randomForest")
set.seed(59)

apartments_rf_model <- randomForest(m2.price ~ construction.year + surface + floor +
      no.rooms + district, data = apartments)

explainer_rf <- explain(apartments_rf_model,
      data = apartmentsTest[,2:6], y = apartmentsTest$m2.price)

apartments_small <- select_sample(apartmentsTest, 10)

cp_rf <- individual_variable_profile(explainer_rf, apartments_small)
cp_rf
}
}
