% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_oscillations.R
\name{individual_variable_oscillations}
\alias{individual_variable_oscillations}
\title{Calculates Oscillations for Individual Variable Profiles}
\usage{
individual_variable_oscillations(x, sort = TRUE, ...)
}
\arguments{
\item{x}{an individual variable profile explainer produced with the `individual_variable_profile()` function}

\item{sort}{a logical value. If `TRUE` then rows are sorted along the oscillations}

\item{...}{other arguments}
}
\description{
Calculates Oscillations for Individual Variable Profiles
}
\examples{
library("DALEX2")
 \dontrun{
library("randomForest")
set.seed(59)

apartments_rf_model <- randomForest(m2.price ~ construction.year + surface + floor +
      no.rooms + district, data = apartments)

explainer_rf <- explain(apartments_rf_model,
      data = apartments_test, y = apartments_test$m2.price)

apartment <- apartments_test[1,]

ivp_rf <- individual_variable_profile(explainer_rf, apartment)
individual_variable_oscillations(ivp_rf)
}
}
